Updated Code

+++ eForms/Pages/Admin/Settings/AnnexBuildings/AnnexBuildings.cshtml.cs
@@ -6,6 +6,7 @@ using eForms.Services.Interfaces;
 using eForms.Services.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using eForms.Domain.Enums;
 
 namespace eForms.Pages.Admin
 {
@@ -14,10 +15,12 @@ namespace eForms.Pages.Admin
         public IEnumerable<BuildingAnnexModel> buildings { get; set; }
 
         IBuildingService buildingService;
+        IAuthService authService;
 
-        public AnnexBuildingsModel(IBuildingService _buildingService)
+        public AnnexBuildingsModel(IBuildingService _buildingService, IAuthService _authService)
         {
             buildingService = _buildingService;
+            authService = _authService;
         }
         public async Task OnGet()
         {
@@ -25,8 +28,15 @@ namespace eForms.Pages.Admin
         }
         public async Task<IActionResult> OnPostDelete(int id)
         {
-            await buildingService.DeleteBuilding(id);
-            return RedirectToPage("AnnexBuildings");
+            if (await authService.Check(Roles.Admin, 0) == false)
+            {
+                return RedirectToPage("/Error", "Error", new { @Id = 401 });
+            }
+            else
+            {
+                await buildingService.DeleteBuilding(id);
+                return RedirectToPage("AnnexBuildings");
+            }
         }
     }
 }
