+++ eForms/Pages/Admin/Settings/Countries/Countries.cshtml.cs
@@ -1,7 +1,9 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Threading.Tasks;
+using eForms.Domain.Enums;
 using eForms.Services.Interfaces;
 using eForms.Services.Models;
 using Microsoft.AspNetCore.Mvc;
@@ -14,10 +16,11 @@ namespace eForms.Pages.Admin
         public IEnumerable<CountryModel> countries { get; set; }
 
         ICountryService countryService;
-
-        public CountriesModel(ICountryService _countryService)
+        IAuthService authService;
+        public CountriesModel(ICountryService _countryService, IAuthService _authService)
         {
             countryService = _countryService;
+            authService = _authService;
         }
         public async Task OnGet()
         {
@@ -25,8 +28,16 @@ namespace eForms.Pages.Admin
         }
         public async Task<IActionResult> OnPostDelete(int id)
         {
-            await countryService.DeleteCountry(id);
-            return RedirectToPage("Countries");
+            if (await authService.Check(Roles.Admin, 0) == false)
+            {
+                return RedirectToPage("/Error", "Error", new { @Id = 401 });
+            }
+            else
+            {
+                await countryService.DeleteCountry(id);
+                return RedirectToPage("Countries");
+            }
+            
         }
     }
 }
